+++ packages/backend/src/models/ProjectModel/ProjectModel.ts
@@ -1544,17 +1544,19 @@ export class ProjectModel {
                     ExploreType.VIRTUAL,
                 );
 
-            Logger.info(
-                `Duplicating ${virtualViews.length} virtual views into ${previewProjectUuid}`,
-            );
+            if (virtualViews.length > 0) {
+                Logger.info(
+                    `Duplicating ${virtualViews.length} virtual views into ${previewProjectUuid}`,
+                );
 
-            await trx(CachedExploreTableName).insert(
-                virtualViews.map((v) => ({
-                    ...v,
-                    project_uuid: previewProjectUuid,
-                    cached_explore_uuid: undefined,
-                })),
-            );
+                await trx(CachedExploreTableName).insert(
+                    virtualViews.map((v) => ({
+                        ...v,
+                        project_uuid: previewProjectUuid,
+                        cached_explore_uuid: undefined,
+                    })),
+                );
+            }
 
             // .dP"Y8    db    Yb    dP 888888 8888b.      .dP"Y8  dP"Yb  88
             // `Ybo."   dPYb    Yb  dP  88__    8I  Yb     `Ybo." dP   Yb 88